HSEARCH-4620 Hide sync strategies implementations
