+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/plan/synchronization/impl/ReadSyncIndexingPlanSynchronizationStrategy.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.plan.synchronization.spi;
+package org.hibernate.search.mapper.pojo.plan.synchronization.impl;
 
 import java.lang.invoke.MethodHandles;
 
@@ -19,27 +19,26 @@ import org.hibernate.search.util.common.impl.Futures;
 import org.hibernate.search.util.common.logging.impl.LoggerFactory;
 
 @Incubating
-public final class SyncIndexingPlanSynchronizationStrategy
-		implements IndexingPlanSynchronizationStrategy {
+public final class ReadSyncIndexingPlanSynchronizationStrategy implements IndexingPlanSynchronizationStrategy {
 
-	public static final IndexingPlanSynchronizationStrategy INSTANCE = new SyncIndexingPlanSynchronizationStrategy();
+	public static final IndexingPlanSynchronizationStrategy INSTANCE = new ReadSyncIndexingPlanSynchronizationStrategy();
 	private static final Log log = LoggerFactory.make( Log.class, MethodHandles.lookup() );
 
-	private SyncIndexingPlanSynchronizationStrategy() {
+	private ReadSyncIndexingPlanSynchronizationStrategy() {
 	}
 
 	@Override
 	public String toString() {
-		return IndexingPlanSynchronizationStrategy.class.getSimpleName() + ".sync()";
+		return IndexingPlanSynchronizationStrategy.class.getSimpleName() + ".readSync()";
 	}
 
 	@Override
 	public void apply(IndexingPlanSynchronizationStrategyConfigurationContext context) {
-		// Request indexing to force a commit and a refresh.
-		context.documentCommitStrategy( DocumentCommitStrategy.FORCE );
+		// Request indexing to force a refresh, but not necessarily a commit.
+		context.documentCommitStrategy( DocumentCommitStrategy.NONE );
 		context.documentRefreshStrategy( DocumentRefreshStrategy.FORCE );
 		context.indexingFutureHandler( future -> {
-			// Wait for the result of indexing, so that we're sure changes were committed and refreshed.
+			// Wait for the result of indexing, so that we're sure changes were applied and refreshed.
 			SearchIndexingPlanExecutionReport report = Futures.unwrappedExceptionJoin( future );
 			report.throwable().ifPresent( t -> {
 				throw log.indexingFailure( t.getMessage(), report.failingEntities(), t );
